Separate two implementations defined in same file
